// src/theme/tokens-import.d.ts
export declare const lightTokens: {
    colors: {
        primary: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        secondary: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        error: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        warning: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        info: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        success: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        background: {
            default: string;
            paper: string;
        };
        text: {
            primary: string;
            secondary: string;
            disabled: string;
        };
        action: {
            active: string;
            hover: string;
            selected: string;
            disabled: string;
            disabledBackground: string;
        };
        divider: string;
    };
    components: {
        alert: {
            borderRadius: string;
            error: {
                filled: {
                    background: string;
                    foreground: string;
                };
                outlined: {
                    background: string;
                    foreground: string;
                    border: string;
                };
                standard: {
                    background: string;
                    foreground: string;
                };
            };
            warning: {
                filled: {
                    background: string;
                    foreground: string;
                };
                outlined: {
                    background: string;
                    foreground: string;
                    border: string;
                };
                standard: {
                    background: string;
                    foreground: string;
                };
            };
            info: {
                filled: {
                    background: string;
                    foreground: string;
                };
                outlined: {
                    background: string;
                    foreground: string;
                    border: string;
                };
                standard: {
                    background: string;
                    foreground: string;
                };
            };
            success: {
                filled: {
                    background: string;
                    foreground: string;
                };
                outlined: {
                    background: string;
                    foreground: string;
                    border: string;
                };
                standard: {
                    background: string;
                    foreground: string;
                };
            };
        };
        input: {
            borderRadius: string;
            standard: {
                enabledBorder: string;
            };
            filled: {
                enabledFill: string;
            };
            outlined: {
                enabledBorder: string;
            };
        };
        paper: {
            elevation0: string;
            elevation1: string;
            elevation2: string;
        };
        chip: {
            defaultCloseFill: string;
            defaultEnabledBorder: string;
        };
        button: {
            borderRadius: string;
        };
        card: {
            borderRadius: string;
        };
        dialog: {
            borderRadius: string;
        };
        table: {
            border: string;
            header: {
                background: string;
            };
        };
        rating: {
            activeFill: string;
            enabledBorder: string;
        };
    };
};
export declare const darkTokens: {
    colors: {
        primary: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        secondary: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        error: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        warning: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        info: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        success: {
            main: string;
            dark: string;
            light: string;
            contrastText: string;
        };
        background: {
            default: string;
            paper: string;
        };
        text: {
            primary: string;
            secondary: string;
            disabled: string;
        };
        action: {
            active: string;
            hover: string;
            selected: string;
            disabled: string;
            disabledBackground: string;
        };
        divider: string;
    };
    components: {
        alert: {
            borderRadius: string;
            error: {
                filled: {
                    background: string;
                    foreground: string;
                };
                outlined: {
                    background: string;
                    foreground: string;
                    border: string;
                };
                standard: {
                    background: string;
                    foreground: string;
                };
            };
            warning: {
                filled: {
                    background: string;
                    foreground: string;
                };
                outlined: {
                    background: string;
                    foreground: string;
                    border: string;
                };
                standard: {
                    background: string;
                    foreground: string;
                };
            };
            info: {
                filled: {
                    background: string;
                    foreground: string;
                };
                outlined: {
                    background: string;
                    foreground: string;
                    border: string;
                };
                standard: {
                    background: string;
                    foreground: string;
                };
            };
            success: {
                filled: {
                    background: string;
                    foreground: string;
                };
                outlined: {
                    background: string;
                    foreground: string;
                    border: string;
                };
                standard: {
                    background: string;
                    foreground: string;
                };
            };
        };
        input: {
            borderRadius: string;
            standard: {
                enabledBorder: string;
            };
            filled: {
                enabledFill: string;
            };
            outlined: {
                enabledBorder: string;
            };
        };
        paper: {
            elevation0: string;
            elevation1: string;
            elevation2: string;
            elevation4: string;
            elevation6: string;
            elevation8: string;
            elevation12: string;
            elevation16: string;
            elevation24: string;
        };
        chip: {
            defaultCloseFill: string;
            defaultEnabledBorder: string;
        };
        button: {
            borderRadius: string;
        };
        card: {
            borderRadius: string;
        };
        dialog: {
            borderRadius: string;
        };
        table: {
            border: string;
            header: {
                background: string;
            };
        };
        rating: {
            activeFill: string;
            enabledBorder: string;
        };
    };
};